spark/requirements.txt
pyspark
python-dotenv
google-cloud-bigquery
